pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["SizNi <wwwwwwwww@list.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8.1"
flask = "^2.2.2"
gunicorn = "^20.1.0"
flake8 = "^6.0.0"
pytest = "^7.2.0"
python-dotenv = "^0.21.0"
flask-wtf = "^1.0.1"
railway = "^0.0.4"
pip = "^22.3.1"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"



[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
